chore: move types to definition folder and fix imports
